+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 Adjustment starts at 0% at zoom 6 and reaches 100% at zoom 9, configured by `minAdjustZoom` and `maxAdjustZoom` options.
 - Added `Map` `numWorkers` option (7 by default).
 - Added default `Map` `center` and `zoom` (`[0, 0], 0`).
+- Removed `Map` `getUUID` method, added `util` `uniqueId()` instead.
 
 ### 0.0.10